+++ include/telebot/format.h
@@ -1,0 +1,10 @@
+#ifndef FORMAT_H
+#define FORMAT_H
+
+#include <telebot/telebot.h>
+
+char *format(char *format, ...);
+char *vsformat (char *formats, va_list params);
+char *remove_newline(char *str);
+
+#endif
