+++ include/telebot/memory.h
@@ -1,0 +1,9 @@
+#ifndef MEMORY_H
+#define MEMORY_H
+
+#include <telebot/telebot.h>
+
+char *alloc_string(const char * str);
+char *realloc_string(char *base, char *str);
+
+#endif
