+++ lib/objects.c
@@ -22,7 +22,7 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 */
 
-#include <telebot.h>
+#include <telebot/telebot.h>
 
 static Error *_error = NULL;
