+++ lib/telebot.c
@@ -1,4 +1,4 @@
-#include <telebot.h>
+#include <telebot/telebot.h>
 
 /* parameter parse_mode */
 #define MODEHTML "HTML"
