+++ lib/util.c
@@ -1,4 +1,4 @@
-#include <util.h>
+#include <telebot/util.h>
 
 // format = (char *param_alias), (bool param), (char *param_alias), (bool param), ...
 char *vsboolean_param_parser (char *base, int psize, ...) {
